AKI-332: IExternalState returns block difficulty as BigInteger
- This is to account for the fact that the block difficulty is a 16-byte
  quantity (even though in practice we have never seen it remotely close
  to this limit).

+++ org.aion.avm.core/src/org/aion/avm/core/IExternalState.java
@@ -283,7 +283,7 @@ public interface IExternalState {
      *
      * @return the current best block's difficulty.
      */
-    public long getBlockDifficulty();
+    public BigInteger getBlockDifficulty();
 
     /**
      * Returns the address of the miner.
