+++ org.aion.avm.core/src/org/aion/kernel/TestingState.java
@@ -254,8 +254,8 @@ public class TestingState implements IExternalState {
     }
 
     @Override
-    public long getBlockDifficulty() {
-        return blockDifficulty.longValue();
+    public BigInteger getBlockDifficulty() {
+        return blockDifficulty;
     }
 
     @Override
